Decouple harvest model from the rest of the plugin
Separate HarvestMetadataAPIPlugin to its own module, update imports on
CLI commands. This CLI would be better off in a separate plugin but we
don't want to add more complexity and keep compatibility
